Steganography technique.
Steganography is the practice of concealing a file, message, image, or video within another file, message, image, or video. And then it should display as follow sequence way. The advantage of steganography over cryptography alone is that the intended secret message does not attract attention to itself as an object of scrutiny.