The Tipping Problem - The Hard Way
Note: This method computes everything by hand, step by step. For most people, the new API for fuzzy systems will be preferable. The same problem is solved with the new API in this example.
The ‘tipping problem’ is commonly used to illustrate the power of fuzzy logic principles to generate complex behavior from a compact, intuitive set of expert rules.
Input variables
A number of variables play into the decision about how much to tip while dining. Consider two of them:

quality : Quality of the food
service : Quality of the service
Output variable
The output variable is simply the tip amount, in percentage points:

tip : Percent of bill to add as tip
For the purposes of discussion, let’s say we need ‘high’, ‘medium’, and ‘low’ membership functions for both input variables and our output variable. These are defined in scikit-fuzzy as follows